Move over specs from ProofDocument* classes
